Updated the home page to show when someone is an invited guest
